+++ packages/backend/src/models/UserModel.ts
@@ -25,6 +25,7 @@ import {
 } from '@lightdash/common';
 import bcrypt from 'bcrypt';
 import { Knex } from 'knex';
+import NodeCache from 'node-cache';
 import { LightdashConfig } from '../config/parseConfig';
 import {
     createEmail,
@@ -120,13 +121,45 @@ export class UserModel {
 
     private readonly database: Knex;
 
+    private readonly sessionUserCache: NodeCache | undefined;
+
     constructor({ database, lightdashConfig }: UserModelArguments) {
         this.database = database;
         this.lightdashConfig = lightdashConfig;
+
+        // Initialize cache with 30 seconds TTL
+        this.sessionUserCache =
+            process.env.EXPERIMENTAL_CACHE === 'true'
+                ? new NodeCache({
+                      stdTTL: 30, // time to live in seconds
+                      checkperiod: 60, // cleanup interval in seconds
+                  })
+                : undefined;
     }
 
     private canTrackingBeAnonymized() {
         return this.lightdashConfig.mode !== LightdashMode.CLOUD_BETA;
+    }
+
+    async getSessionUserFromCacheOrDB(
+        userUuid: string,
+        organizationUuid: string,
+    ) {
+        const cacheKey = `${userUuid}::${organizationUuid}`;
+        // Try to get from cache first
+        const cachedUser = this.sessionUserCache?.get<SessionUser>(cacheKey);
+        if (cachedUser) {
+            // Return cached user
+            return { sessionUser: cachedUser, cacheHit: true };
+        }
+        // If not in cache, get from database
+        const sessionUser = await this.findSessionUserAndOrgByUuid(
+            userUuid,
+            organizationUuid,
+        );
+        // Store in cache
+        this.sessionUserCache?.set(cacheKey, sessionUser);
+        return { sessionUser, cacheHit: false };
     }
 
     // DB Errors:
@@ -948,12 +981,12 @@ export class UserModel {
 
             const projectMemberships = Object.entries(projects || {}).map(
                 async ([projectUuid, projectRole]) => {
-                    const [project] = await this.database('projects')
+                    const [project] = await trx('projects')
                         .select('project_id')
                         .where('project_uuid', projectUuid);
 
                     if (project) {
-                        await this.database('project_memberships').insert({
+                        await trx('project_memberships').insert({
                             project_id: project.project_id,
                             role: projectRole,
                             user_id: user.user_id,