+++ packages/mantine-v7/src/index.ts
@@ -1,0 +1,9 @@
+import '@mantine/core/styles.css';
+
+export {
+    MantineProvider,
+    Tree,
+    getTreeExpandedState,
+    useTree,
+    type RenderTreeNodePayload,
+} from '@mantine/core';